# labs/v0.9-schema-typed-probes/openapi.yaml
openapi: 3.1.0
info:
  title: ParamIntel v0.9 Schema-Typed Probe Acceptance Lab
  version: 1.0.0
paths:
  /boolean-real:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                profile:
                  type: object
                  properties:
                    name:
                      type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile:
                    type: object
                    properties:
                      name:
                        type: string
                      beta_access:
                        type: boolean
                        readOnly: true
  /boolean-noise:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                profile:
                  type: object
                  properties:
                    name:
                      type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile:
                    type: object
                    properties:
                      name:
                        type: string
                      beta_access:
                        type: boolean
                        readOnly: true
  /integer-real:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                profile:
                  type: object
                  properties:
                    name:
                      type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile:
                    type: object
                    properties:
                      name:
                        type: string
                      access_level:
                        type: integer
                        readOnly: true
  /union:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                profile:
                  type: object
                  properties:
                    name:
                      type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile:
                    type: object
                    properties:
                      name:
                        type: string
                      beta_access:
                        type:
                          - boolean
                          - 'null'
                        readOnly: true